Extracted template from menu component
