+++ src/components/menu/index.js
@@ -1,0 +1,8 @@
+import Component from "../component";
+import {template} from "./template";
+
+export default class Menu extends Component {
+  constructor(films) {
+    super(template(films));
+  }
+}
